requirements.txt
accelerate==1.10.1
altair==5.5.0
annotated-types==0.7.0
anthropic==0.64.0
anyio==4.10.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
bitsandbytes==0.42.0
blinker==1.9.0
build==1.3.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
chromadb==1.0.20
click==8.2.1
coloredlogs==15.0.1
cryptography==45.0.6
distro==1.9.0
durationpy==0.10
fastapi==0.116.1
filelock==3.19.1
flatbuffers==25.2.10
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.1.9
httpcore==1.0.9
httplib2==0.30.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jpype1==1.6.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kubernetes==33.1.0
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
narwhals==2.2.0
networkx==3.5
numpy==2.2.6
oauthlib==3.3.1
-e git+https://github.com/JaynouOliver/oct.git@89655f6cdf4746609f6588acf677e507f1804d80#egg=od_parse&subdirectory=od-parse
onnxruntime==1.22.1
openai==1.102.0
opencv-python==4.12.0.88
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.3
overrides==7.7.0
packaging==25.0
pandas==2.3.2
pdf2image==1.17.0
pdfminer.six==20250506
pillow==11.3.0
posthog==5.4.0
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
PyMuPDF==1.26.4
pyparsing==3.2.3
PyPika==0.48.9
pyproject_hooks==1.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.8.29
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
sentence-transformers==5.1.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.47.3
streamlit==1.49.1
sympy==1.14.0
tabula-py==2.10.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.0
toml==0.10.2
torch==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.56.0
typer==0.17.3
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
zipp==3.23.0
